fix issues with merging fragments

+++ fragmentmerge-maven-plugin/src/main/java/org/jaffa/plugins/FragmentMergeMojo.java
@@ -54,6 +54,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import org.apache.maven.plugin.AbstractMojo;
@@ -104,7 +105,35 @@ public class FragmentMergeMojo extends AbstractMojo{
      */
     File targetDirectory;
 
+    /**
+     * Skip Tags For configuration files
+     */
+    @Parameter
+    String[] skipTagForConfigFiles;
 
+    /**
+     * Skip Listed configuration files
+     */
+    @Parameter
+    String[] skipConfigFiles;
+
+
+    /**
+     * Getter for skipTagForConfigFiles parameter
+     * @return
+     */
+    public List<String> getSkipTagForConfigFilesList(){
+        return skipTagForConfigFiles !=null ? Arrays.asList(skipTagForConfigFiles) : new ArrayList<String>();
+    }
+
+
+    /**
+     * Getter for skipConfigFiles parameter
+     * @return
+     */
+    public List<String> getSkipConfigFilesList(){
+        return skipConfigFiles !=null ? Arrays.asList(skipConfigFiles) : new ArrayList<String>();
+    }
 
     /**
      * Execute method for Fragment Merging
@@ -128,49 +157,67 @@ public class FragmentMergeMojo extends AbstractMojo{
             if (targetDirectory.exists()) {
 
                 //Merging ApplicationResourceFragments
-                mergeApplicationResources(applicationResources);
-
-                //Merging Dwr Resource Fragments
-                mergeDwrResources(dwr);
-
-                //Merging jawr Resource Fragments
-                mergeJawrResources(jawr);
-
-                //Start Merging Struts-Config
-                boolean strutsConfigFragsFound = false;
-                Fragments.writeTag(strutsConfig, STRUTS_CONFIG_START_TAG);
-                //Merging StrutsFormBean
-                strutsConfigFragsFound = mergeStrutsFormResources(strutsConfig);
-
-                //Merging StrutsGlobalForward
-                strutsConfigFragsFound |= mergeStrutsGlobalForwardResources(strutsConfig);
-
-                //Merging StrutsConfigAction
-                strutsConfigFragsFound |= mergeStrutsConfigActionResources(strutsConfig);
-
-                //End Merging Struts-Config
-                Fragments.writeTag(strutsConfig, STRUTS_CONFIG_END_TAG);
-                if(!strutsConfigFragsFound){
-                    strutsConfig.delete();
+                if(!getSkipConfigFilesList().contains(APPLICATION_RESOURCES)) {
+                    mergeApplicationResources(applicationResources);
                 }
 
-                //Merging tiles-def
-                mergeTileDefsResources(tilesDef);
+                //Merging Dwr Resource Fragments
+                if(!getSkipConfigFilesList().contains(DWR)) {
+                    mergeDwrResources(dwr);
+                }
 
-                //Merging componentDefinitions
-                mergeComponentsResources(components);
+                //Merging jawr Resource Fragments
+                if(!getSkipConfigFilesList().contains(JAWR)) {
+                    mergeJawrResources(jawr);
+                }
 
-                //Merge ApplicationRules
-                mergeApplicationRules();
+                //Start Merging Struts-Config
+                if(!getSkipConfigFilesList().contains(STRUTS_CONFIG)) {
+                    boolean strutsConfigFragsFound = false;
+                    Fragments.writeTag(strutsConfig, STRUTS_CONFIG_START_TAG);
+                    //Merging StrutsFormBean
+                    strutsConfigFragsFound = mergeStrutsFormResources(strutsConfig);
 
-                //Merge business-functions
-                mergeBusinessFunctions();
+                    //Merging StrutsGlobalForward
+                    strutsConfigFragsFound |= mergeStrutsGlobalForwardResources(strutsConfig);
 
-                //Merge roles
-                mergeRoles();
+                    //Merging StrutsConfigAction
+                    strutsConfigFragsFound |= mergeStrutsConfigActionResources(strutsConfig);
 
-                //Merge Navigation
-                mergeNavigation();
+                    //End Merging Struts-Config
+                    Fragments.writeTag(strutsConfig, STRUTS_CONFIG_END_TAG);
+
+                    if(!strutsConfigFragsFound){
+                        strutsConfig.delete();
+                    }
+                    //Merging tiles-def
+                    mergeTileDefsResources(tilesDef);
+                }
+
+                if(!getSkipConfigFilesList().contains(COMPONENTS)) {
+                    //Merging componentDefinitions
+                    mergeComponentsResources(components);
+                }
+
+                if(!getSkipConfigFilesList().contains(APPLICATION_RULES)) {
+                    //Merge ApplicationRules
+                    mergeApplicationRules();
+                }
+
+                if(!getSkipConfigFilesList().contains(BUSINESS_FUNCTIONS)) {
+                    //Merge business-functions
+                    mergeBusinessFunctions();
+                }
+
+                if(!getSkipConfigFilesList().contains(ROLES)) {
+                    //Merge roles
+                    mergeRoles();
+                }
+
+                if(!getSkipConfigFilesList().contains(NAVIGATION)) {
+                    //Merge Navigation
+                    mergeNavigation();
+                }
             }
         }catch(IOException io){
             getLog().error(io);
@@ -184,7 +231,7 @@ public class FragmentMergeMojo extends AbstractMojo{
         FileFinder appResourcesFinder = new FileFinder("*"+APPLICATION_RESOURCES+"*."+PFRAGMENT);
         Files.walkFileTree(targetDirectory.toPath(), appResourcesFinder);
         List<Path> applicationResourceFragFiles = appResourcesFinder.getFiles();
-        Fragments.mergeFragmentResources(applicationResources, applicationResourceFragFiles, APP_RESOURCES_START_TAG, APP_RESOURCES_END_TAG);
+        mergeFragmentResources(applicationResources, applicationResourceFragFiles, APP_RESOURCES_START_TAG, APP_RESOURCES_END_TAG);
         getLog().debug("End of ApplicationResources Merge Process");
     }
 
@@ -193,7 +240,7 @@ public class FragmentMergeMojo extends AbstractMojo{
         FileFinder dwrResourcesFinder = new FileFinder(DWR+"*."+XFRAGMENT);
         Files.walkFileTree(targetDirectory.toPath(), dwrResourcesFinder);
         List<Path> dwrResourceFragFiles = dwrResourcesFinder.getFiles();
-        Fragments.mergeFragmentResources(dwr, dwrResourceFragFiles, DWR_START_TAG, DWR_END_TAG);
+        mergeFragmentResources(dwr, dwrResourceFragFiles, DWR_START_TAG, DWR_END_TAG, getSkipTagForConfigFilesList().contains(DWR), true);
         getLog().debug("End of ApplicationResources Merge Process");
     }
 
@@ -204,7 +251,7 @@ public class FragmentMergeMojo extends AbstractMojo{
         if(htmlDirectory!=null) {
             Files.walkFileTree(htmlDirectory, jawrResourceFinder);
             List<Path> jawrResourceFragFiles = jawrResourceFinder.getFiles();
-            Fragments.mergeFragmentResources(jawr, jawrResourceFragFiles, JAWR_START_TAG, JAWR_END_TAG, false);
+            mergeFragmentResources(jawr, jawrResourceFragFiles, JAWR_START_TAG, JAWR_END_TAG, false, false);
         }
         getLog().debug("End of ApplicationResources Merge Process");
     }
@@ -214,7 +261,7 @@ public class FragmentMergeMojo extends AbstractMojo{
         FileFinder strutsConfigFormBeanResourceFinder = new FileFinder(STRUTS_CONFIG_FORM_BEAN+"*."+XFRAGMENT);
         Files.walkFileTree(targetDirectory.toPath(), strutsConfigFormBeanResourceFinder);
         List<Path> strutsConfigFormBeanResourceFragFiles = strutsConfigFormBeanResourceFinder.getFiles();
-        Fragments.mergeFragmentResources(strutsConfig, strutsConfigFormBeanResourceFragFiles, STRUTS_FORM_START_TAG, STRUTS_FORM_END_TAG);
+        mergeFragmentResources(strutsConfig, strutsConfigFormBeanResourceFragFiles, STRUTS_FORM_START_TAG, STRUTS_FORM_END_TAG);
         getLog().debug("End of StrutsFormResources Merge Process");
         return strutsConfigFormBeanResourceFragFiles!=null && strutsConfigFormBeanResourceFragFiles.size() > 0;
     }
@@ -224,7 +271,7 @@ public class FragmentMergeMojo extends AbstractMojo{
         FileFinder strutsConfigGlobalForwardResourceFinder = new FileFinder(STRUTS_CONFIG_GLOBAL_FORWARD+"*."+XFRAGMENT);
         Files.walkFileTree(targetDirectory.toPath(), strutsConfigGlobalForwardResourceFinder);
         List<Path> strutsConfigGlobalForwardResourceFragFiles = strutsConfigGlobalForwardResourceFinder.getFiles();
-        Fragments.mergeFragmentResources(strutsConfig, strutsConfigGlobalForwardResourceFragFiles, STRUTS_GLOBAL_FWD_START_TAG, STRUTS_GLOBAL_FWD_END_TAG);
+        mergeFragmentResources(strutsConfig, strutsConfigGlobalForwardResourceFragFiles, STRUTS_GLOBAL_FWD_START_TAG, STRUTS_GLOBAL_FWD_END_TAG);
         getLog().debug("End of StrutsGlobalForward Merge Process");
         return strutsConfigGlobalForwardResourceFragFiles!=null && strutsConfigGlobalForwardResourceFragFiles.size() > 0;
     }
@@ -234,7 +281,7 @@ public class FragmentMergeMojo extends AbstractMojo{
         FileFinder strutsConfigActionResourceResourceFinder = new FileFinder(STRUTS_CONFIG_ACTION+"*."+XFRAGMENT);
         Files.walkFileTree(targetDirectory.toPath(), strutsConfigActionResourceResourceFinder);
         List<Path> strutsConfigActionResourceFragFiles = strutsConfigActionResourceResourceFinder.getFiles();
-        Fragments.mergeFragmentResources(strutsConfig, strutsConfigActionResourceFragFiles, STRUTS_ACTION_START_TAG, STRUTS_ACTION_END_TAG);
+        mergeFragmentResources(strutsConfig, strutsConfigActionResourceFragFiles, STRUTS_ACTION_START_TAG, STRUTS_ACTION_END_TAG);
         getLog().debug("End of StrutsConfig Merge Process");
         return strutsConfigActionResourceFragFiles!=null && strutsConfigActionResourceFragFiles.size() > 0;
     }
@@ -244,7 +291,7 @@ public class FragmentMergeMojo extends AbstractMojo{
         FileFinder componentTilesDefinitionsResourceFinder = new FileFinder(TILE_DEFS+"*."+XFRAGMENT);
         Files.walkFileTree(targetDirectory.toPath(), componentTilesDefinitionsResourceFinder);
         List<Path> componentTilesDefinitionsResourceFragFiles = componentTilesDefinitionsResourceFinder.getFiles();
-        Fragments.mergeFragmentResources(tilesDef, componentTilesDefinitionsResourceFragFiles, STRUTS_TILE_DEFS_START_TAG, STRUTS_TILE_DEFS_END_TAG);
+        mergeFragmentResources(tilesDef, componentTilesDefinitionsResourceFragFiles, STRUTS_TILE_DEFS_START_TAG, STRUTS_TILE_DEFS_END_TAG);
         getLog().debug("End of TilesDef Merge Process");
     }
 
@@ -253,14 +300,14 @@ public class FragmentMergeMojo extends AbstractMojo{
         FileFinder componentDefinitionsResourceFinder = new FileFinder(COMPONENT_DEFINITIONS+"*."+XFRAGMENT);
         Files.walkFileTree(targetDirectory.toPath(), componentDefinitionsResourceFinder);
         List<Path> componentTilesDefinitionsResourceFragFiles = componentDefinitionsResourceFinder.getFiles();
-        Fragments.mergeFragmentResources(components, componentTilesDefinitionsResourceFragFiles, COMPONENTS_START_TAG, COMPONENTS_END_TAG);
+        mergeFragmentResources(components, componentTilesDefinitionsResourceFragFiles, COMPONENTS_START_TAG, COMPONENTS_END_TAG);
         getLog().debug("End of Components Merge Process");
     }
 
 
     private void mergeApplicationRules() throws IOException {
         getLog().debug("Starting ApplicationRules Merge Process");
-        FileFinder applicationRulesResourceFinder = new FileFinder(APPLICATION_RULES+"*."+PROPERTIES);
+        FileFinder applicationRulesResourceFinder = new FileFinder(APPLICATION_RULES_ +"*."+PROPERTIES);
         Files.walkFileTree(targetDirectory.toPath(), applicationRulesResourceFinder);
         List<Path> applicationRulesFiles = applicationRulesResourceFinder.getFiles();
         if(applicationRulesFiles!=null && applicationRulesFiles.size() > 0) {
@@ -268,7 +315,7 @@ public class FragmentMergeMojo extends AbstractMojo{
                 File mergedApplicationRule = new File(classesDirectory + META_INF_LOCATION + applicationRulesFile.getFileName());
                 List<Path> applicationRuleFilesList = new ArrayList<>();
                 applicationRuleFilesList.add(applicationRulesFile);
-                Fragments.mergeFragmentResources(mergedApplicationRule, applicationRuleFilesList, EMPTY_START_TAG, EMPTY_END_TAG);
+                mergeFragmentResources(mergedApplicationRule, applicationRuleFilesList, EMPTY_START_TAG, EMPTY_END_TAG);
             }
         }
         getLog().debug("End of ApplicationResources Merge Process");
@@ -284,7 +331,7 @@ public class FragmentMergeMojo extends AbstractMojo{
                 File mergedBusinessFunctionFile = new File(classesDirectory + META_INF_LOCATION + businessFunctionFile.getFileName());
                 List<Path> businessFunctionFilesList = new ArrayList<>();
                 businessFunctionFilesList.add(businessFunctionFile);
-                Fragments.mergeFragmentResources(mergedBusinessFunctionFile, businessFunctionFilesList, EMPTY_START_TAG, EMPTY_END_TAG);
+                mergeFragmentResources(mergedBusinessFunctionFile, businessFunctionFilesList, EMPTY_START_TAG, EMPTY_END_TAG);
             }
         }
         getLog().debug("End of Business Functions Merge Process");
@@ -300,7 +347,7 @@ public class FragmentMergeMojo extends AbstractMojo{
                 File mergedRoles = new File(classesDirectory + META_INF_LOCATION + rolesFile.getFileName());
                 List<Path> rolesFilesList = new ArrayList<>();
                 rolesFilesList.add(rolesFile);
-                Fragments.mergeFragmentResources(mergedRoles, rolesFilesList, EMPTY_START_TAG, EMPTY_END_TAG);
+                mergeFragmentResources(mergedRoles, rolesFilesList, EMPTY_START_TAG, EMPTY_END_TAG);
             }
         }
         getLog().debug("End of Roles Merge Process");
@@ -316,10 +363,22 @@ public class FragmentMergeMojo extends AbstractMojo{
                 File mergedNavigation = new File(classesDirectory + META_INF_LOCATION + navigationFile.getFileName());
                 List<Path> navigationFilesList = new ArrayList<>();
                 navigationFilesList.add(navigationFile);
-                Fragments.mergeFragmentResources(mergedNavigation, navigationFilesList, EMPTY_START_TAG, EMPTY_END_TAG);
+                mergeFragmentResources(mergedNavigation, navigationFilesList, EMPTY_START_TAG, EMPTY_END_TAG);
             }
         }
         getLog().debug("End of Navigation Merge Process");
+    }
+
+    private void mergeFragmentResources(File mergedFile, List<Path> fragments, String startTag, String endTag) throws IOException {
+        mergeFragmentResources(mergedFile, fragments, startTag, endTag, false, true);
+    }
+
+    private void mergeFragmentResources(File mergedFile, List<Path> fragments, String startTag, String endTag, Boolean skipTags, Boolean deleteFrags) throws IOException {
+        if(skipTags!=null && Boolean.TRUE.equals(skipTags)){
+            Fragments.mergeFragmentResourcesWithNoTags(mergedFile, fragments, deleteFrags);
+        }else{
+            Fragments.mergeFragmentResources(mergedFile, fragments, startTag, endTag, deleteFrags);
+        }
     }
 }
